requirements.txt
psycopg2-binary>=2.9.9
python-telegram-bot>=20.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
requests>=2.31.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
